Conflicts:
	src/main/java/YogaStudio/dao/generic/CustomerDAO.java
	src/main/java/YogaStudio/dao/generic/SectionDAO.java
	src/main/java/YogaStudio/dao/generic/SemesterDAO.java
	src/main/java/YogaStudio/domain/CustomerEntity.java
	src/main/java/YogaStudio/domain/SectionEntity.java
	src/main/java/YogaStudio/service/CustomerService.java
	src/main/java/YogaStudio/service/SemesterService.java
	src/main/java/YogaStudio/service/UserService.java
	src/main/webapp/WEB-INF/rootconfig.xml
	src/main/webapp/WEB-INF/views/classes/classDetail.jsp
	src/main/webapp/WEB-INF/views/section/editSection.jsp
	src/main/webapp/WEB-INF/views/section/semesterPop.jsp
	src/main/webapp/WEB-INF/views/waiver/requestWaiver.jsp

+++ src/main/java/YogaStudio/Controller/UserController.java
@@ -147,8 +147,8 @@ public class UserController {
     }
     
     @RequestMapping(value = "/user/users/delete", method = RequestMethod.POST)
-    public String delete(int userId) {
-        userService.delete(Long.valueOf(userId));
+    public String delete(Long userId) {
+        userService.delete(userId);
         return "redirect:/users";
     }
